GeometryOps

Documentation for GeometryOps.

GeometryOps.DouglasPeucker

GeometryOps.RadialDistance

GeometryOps.VisvalingamWhyatt

GeometryOps._distance

GeometryOps.apply

GeometryOps.flatten

GeometryOps.flip

GeometryOps.intersection

GeometryOps.intersects

GeometryOps.isclockwise

GeometryOps.isconcave

GeometryOps.point_in_polygon

GeometryOps.point_on_line

GeometryOps.rebuild

GeometryOps.reconstruct

GeometryOps.reproject

GeometryOps.signed_area

GeometryOps.signed_distance

GeometryOps.simplify

GeometryOps.unwrap

GeometryOps.DouglasPeucker — Type

DouglasPeucker <: SimplifyAlg

DouglasPeucker(; number, ratio, tol)

Simplifies geometries by removing points below tol distance from the line between its neighboring points.

Keywords

ratio: the fraction of points that should remain after simplify. Useful as it will generalise for large collections of objects.

number: the number of points that should remain after simplify. Less useful for large collections of mixed size objects.

tol: the minimum distance a point will be from the line joining its neighboring points.

GeometryOps.RadialDistance — Type

RadialDistance <: SimplifyAlg

Simplifies geometries by removing points less than tol distance from the line between its neighboring points.

Keywords

ratio: the fraction of points that should remain after simplify. Useful as it will generalise for large collections of objects.

number: the number of points that should remain after simplify. Less useful for large collections of mixed size objects.

tol: the minimum distance between points.

GeometryOps.VisvalingamWhyatt — Type

VisvalingamWhyatt <: SimplifyAlg

VisvalingamWhyatt(; kw...)

Simplifies geometries by removing points below tol distance from the line between its neighboring points.

Keywords

ratio: the fraction of points that should remain after simplify. Useful as it will generalise for large collections of objects.

number: the number of points that should remain after simplify. Less useful for large collections of mixed size objects.

tol: the minimum area of a triangle made with a point and its neighboring points.

GeometryOps._distance — Method

Distance from p0 to the line segment formed by p1 and p2. Implementation from Turf.jl.

GeometryOps.apply — Method

apply(f, target::Type{<:AbstractTrait}, obj; crs)

Reconstruct a geometry or feature using the function f on the target trait.

f(target_geom) => x where x also has the target trait, or an equivalent.

The result is an functionally similar geometry with values depending on f

Flipped point the order in any feature or geometry, or iterables of either:

```juia import GeoInterface as GI import GeometryOps as GO geom = GI.Polygon([GI.LinearRing([(1, 2), (3, 4), (5, 6), (1, 2)]), GI.LinearRing([(3, 4), (5, 6), (6, 7), (3, 4)])])

flipped_geom = GO.apply(GI.PointTrait, geom) do p (GI.y(p), GI.x(p)) end

GeometryOps.flatten — Method

flatten(target::Type{<:GI.AbstractTrait}, geom)

Lazily flatten any geometry, feature or iterator of geometries or features so that objects with the specified trait are returned by the iterator.

GeometryOps.flip — Method

flip(obj)

Swap all of the x and y coordinates in obj, otherwise keeping the original structure (but not necessarily the original type).

GeometryOps.intersection — Method

intersection(line_a, line_b)

Find a point that intersects LineStrings with two coordinates each.

Returns nothing if no point is found.

Examples

julia> 
line1 = LineString([[124.584961,-12.768946],[126.738281,-17.224758]])
LineString(Array{Float64,1}[[124.585, -12.7689], [126.738, -17.2248]])

julia> line2 = LineString([[123.354492,-15.961329],[127.22168,-14.008696]])
LineString(Array{Float64,1}[[123.354, -15.9613], [127.222, -14.0087]])

julia> 
intersection(line1, line2)
Point([125.584, -14.8357])

GeometryOps.intersects — Method

intersects(line_a, line_b)

Check if line_a intersects with line_b.

These can be LineTrait, LineStringTrait or LinearRingTrait

GeometryOps.isclockwise — Method

isclockwise(line::Union{LineString, Vector{Position}})::Bool

Take a ring and return true or false whether or not the ring is clockwise or counter-clockwise.

Examples

import GeoInterface as GI, GeometryOps as GO
julia> 
line = GI.LineString([(0, 0), (1, 1), (1, 0), (0, 0)])
GeoInterface.Wrappers.LineString{false, false, Vector{Tuple{Int64, Int64}}, Nothing, Nothing}([(0, 0), (1, 1), (1, 0), (0, 0)], nothing, nothing)

julia> GO.isclockwise(line)
true

GeometryOps.isconcave — Method

isconcave(poly::Polygon)::Bool

Take a polygon and return true or false as to whether it is concave or not.

Examples

import GeoInterface as GI, GeometryOps as GO
julia> poly = GI.Polygon([[(0, 0), (0, 1), (1, 1), (1, 0), (0, 0)]])
Polygon(Array{Array{Float64,1},1}[[[0.0, 0.0], [0.0, 1.0], [1.0, 1.0], [1.0, 0.0], [0.0, 0.0]]])

julia> GO.isconcave(poly)
false

GeometryOps.point_in_polygon — Function

point_in_polygon(point::Point, polygon::Union{Polygon, MultiPolygon}, ignoreBoundary::Bool=false)::Bool

Take a Point and a Polygon and determine if the point resides inside the polygon. The polygon can be convex or concave. The function accounts for holes.

Examples

import GeoInterface as GI, GeometryOps as GO
julia> point = (-77.0, 44.0)
(-77.0, 44.0)

julia> poly = GI.Polygon([[[-81, 41], [-81, 47], [-72, 47], [-72, 41], [-81, 41]]])
Polygon(Array{Array{Float64,1},1}[[[-81.0, 41.0], [-81.0, 47.0], [-72.0, 47.0], [-72.0, 41.0], [-81.0, 41.0]]])

julia> GO.point_in_polygon(point, poly)
true

GeometryOps.point_on_line — Method

point_on_line(point::Point, line::LineString, ignoreEndVertices::Bool=false)::Bool

Return true if a point is on a line. Accept a optional parameter to ignore the start and end vertices of the linestring.

Examples

import GeoInterface as GI, GeometryOps as GO
julia> point = GI.Point(1, 1)
GeoInterface.Wrappers.Point{false, false, Tuple{Int64, Int64}, Nothing}((1, 1), nothing)

julia> line = GI.LineString([(0, 0), (3, 3), (4, 4)])
GeoInterface.Wrappers.LineString{false, false, Vector{Tuple{Int64, Int64}}, Nothing, Nothing}([(0, 0), (3, 3), (4, 4)], nothing, nothing)

julia> GO.point_on_line(point, line)
true

GeometryOps.rebuild — Method

rebuild(geom, child_geoms)

Rebuild a geometry from child geometries.

By default geometries will be rebuilt as a GeoInterface.Wrappers geometry, but rebuild can have methods added to it to dispatch on geometries from other packages and specify how to rebuild them.

(Maybe it should go into GeoInterface.jl)

GeometryOps.reconstruct — Method

reconstruct(geom, components)

Reconstruct geom from an iterable of component objects that match its structure.

All objects in components must have the same GeoInterface.trait.

Ususally used in combination with flatten.

GeometryOps.reproject — Method

reproject(geometry; source_crs, target_crs, transform, always_xy, time)
reproject(geometry, source_crs, target_crs; always_xy, time)
reproject(geometry, transform; always_xy, time)

Reproject any GeoInterface.jl compatible geometry from source_crs to target_crs.

The returned object will be constructed from GeoInterface.WrapperGeometry geometries, wrapping views of a Vector{Proj.Point{D}}, where D is the dimension.

Arguments

geometry: Any GeoInterface.jl compatible geometries.

source_crs: the source coordinate referece system, as a GeoFormatTypes.jl object or a string.

target_crs: the target coordinate referece system, as a GeoFormatTypes.jl object or a string.

If these a passed as keywords, transform will take priority. Without it target_crs is always needed, and source_crs is needed if it is not retreivable from the geometry with GeoInterface.crs(geometry).

Keywords

-always_xy: force x, y coordinate order, true by default. false will expect and return points in the crs coordinate order. -time: the time for the coordinates. Inf by default.

GeometryOps.signed_area — Method

signed_area(geom)::Real

Returns the signed area of the geometry, based on winding order.

GeometryOps.signed_distance — Method

signed_distance(geom, x::Real, y::Real)::Float64

Calculates the signed distance from the geometry geom to the point defined by (x, y). Points within geom have a negative distance, and points outside of geom have a positive distance.

If geom is a MultiPolygon, then this function returns the maximum distance to any of the polygons in geom.

GeometryOps.simplify — Method

simplify(obj; kw...)
simplify(::SimplifyAlg, obj)

Simplify a geometry, feature, feature collection, or nested vectors or a table of these.

RadialDistance, DouglasPeucker, or VisvalingamWhyatt algorithms are available, listed in order of increasing quality but decreaseing performance.

PoinTrait and MultiPointTrait are returned unchanged.

The default behaviour is simplify(DouglasPeucker(; kw...), obj). Pass in other SimplifyAlg to use other algorithms.

Example

Simplify a polygon to have six points:

import GeoInterface as GI
import GeometryOps as GO

poly = GI.Polygon([[
    [-70.603637, -33.399918],
    [-70.614624, -33.395332],
    [-70.639343, -33.392466],
    [-70.659942, -33.394759],
    [-70.683975, -33.404504],
    [-70.697021, -33.419406],
    [-70.701141, -33.434306],
    [-70.700454, -33.446339],
    [-70.694274, -33.458369],
    [-70.682601, -33.465816],
    [-70.668869, -33.472117],
    [-70.646209, -33.473835],
    [-70.624923, -33.472117],
    [-70.609817, -33.468107],
    [-70.595397, -33.458369],
    [-70.587158, -33.442901],
    [-70.587158, -33.426283],
    [-70.590591, -33.414248],
    [-70.594711, -33.406224],
    [-70.603637, -33.399918]]])

simple = GO.simplify(poly; number=6)
GI.npoint(simple)

# output
6

GeometryOps.unwrap — Function

unwrap(target::Type{<:AbstractTrait}, obj)
unwrap(f, target::Type{<:AbstractTrait}, obj)

Unwrap the geometry to vectors, down to the target trait.

If f is passed in it will be applied to the target geometries as they are found.